requirements.txt
# Core dependencies
python-telegram-bot==20.6
solana==0.30.3
solders>=0.18.1
openai>=1.12.0
aiohttp==3.8.6
asyncio==3.4.3
requests>=2.28.0

# Data handling and visualization
plotly==5.18.0
pandas==2.1.1

# Web server
flask==2.3.3

# Utilities
python-dotenv==1.0.0
PyNaCl>=1.5.0
base58>=2.1.1

# Privy integration
requests>=2.28.0 